+++ packages/eip-5792-middleware/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- **BREAKING:** Use new `Messenger` from `@metamask/messenger` ([#0000](https://github.com/MetaMask/core/pull/0000))
+- **BREAKING:** Use new `Messenger` from `@metamask/messenger` ([#6958](https://github.com/MetaMask/core/pull/6958))
 
 ## [1.2.4]